+++ nginx-proxy-manager/install.sh
@@ -4,16 +4,21 @@ DEFAULT_NEW_PIHOLE_PORT=8000
 
 function confirm() {
     local prompt="$1"
+    local exit_on_no=${2:-false}
     read -p "$prompt [Y/n] " answer
     case "$answer" in
         Y|y|"")
             return 0;;
         N|n)
-            echo "Exiting."
-            exit 1;;
+            if $exit_on_no; then
+                echo "Exit!"
+                exit 1;;
+            else
+                return 1;;
+            fi
         *)
             echo "Invalid response."
-            confirm "$prompt";;
+            return confirm "$prompt" "$exit_on_no";;
     esac
 }
 
@@ -100,7 +105,7 @@ function install_npm() {
     mkdir -p $CONFIG_PATH/data
     mkdir -p $CONFIG_PATH/letsencrypt
 
-    curl -L https://raw.githubusercontent.com/rodneyshupe/pihole-addons/main/nginx-proxy-manager/docker-compose.yml --output $CONFIG_PATH/docker-compose.yml
+    curl -s -L https://raw.githubusercontent.com/rodneyshupe/pihole-addons/main/nginx-proxy-manager/docker-compose.yml --output $CONFIG_PATH/docker-compose.yml
 
     # Create .env file
     echo "APPDATA_ROOT=$HOME/.config" >$CONFIG_PATH/.env
@@ -112,13 +117,14 @@ function install_npm() {
 if ! $(docker-compose -v >/dev/null 2>&1) ; then
     echo "Docker needs to be installed."
     echo ""
-    confirm "Do you want to continue?"
-    install_docker
-    echo ""
-    echo "Reboot required. Once complete rerun the script."
-    echo ""
-    confirm "Do you want to continue?"
-    sudo shutdown --reboot now
+    if confirm "Do you want to continue?"; then
+        install_docker
+        echo ""
+        echo "Reboot required. Once complete rerun the script."
+        echo ""
+        confirm "Do you want to continue?" true
+        sudo shutdown --reboot now
+    fi
 else
     echo "Pi-hole interface needs to be moved to a port other than the default of 80."
 
@@ -126,12 +132,13 @@ else
     echo
 
     echo "About to move Pi-Hole administration to port $port"
-    confirm "Do you want to continue?"
+    if confirm "Do you want to continue?"; then
+        change_lighttpd_config $port
+        echo
 
-    change_lighttpd_config $port
-    echo
-
-    echo "About to install the container for Nginx Proxy Manager"
-    confirm "Do you want to continue?"
-    install_npm
+        echo "About to install the container for Nginx Proxy Manager"
+        if confirm "Do you want to continue?"; then
+            install_npm
+        fi
+    fi
 fi
